Gunnar: -renamed FQHESphereFermionsAllSzToSU2 to FQHESphereWithSpinAllSzToSU2         -created placeholder classes for angular momentum operators in SU4 space         -various small edits
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1628 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/Programs/FQHEOnLattice/FQHELatticeBosonsMeanField.cc
@@ -379,7 +379,7 @@ int main(int argc, char** argv)
 
       File.precision(10);
       File << Energy/NbrSites << "\t" << MeanFieldState.GetNbrParticles()/NbrSites << "\t" << Energy/MeanFieldState.GetNbrParticles() + ChemicalPotential;
-      File.precision(nearbyint(-log(Manager.GetDouble("tolerance"))/log(10)));
+      File.precision(nearbyint(-log(Manager.GetDouble("tolerance"))/log((double)10.0)));
       for (int i=0; i<Parameters.GetVectorDimension(); ++i)
 	File << "\t" << Parameters[i];
       File << endl;
